requirements.txt
pysam
biopython
cython
dendropy
graphviz
matplotlib
numpy
pandas
scipy
seaborn
snakeviz